+++ packages/sv/lib/addons/common.ts
@@ -2,6 +2,7 @@ import { imports, exports, common } from '../core/tooling/js/index.ts';
 import { toFragment, type SvelteAst } from '../core/tooling/svelte/index.ts';
 import { parseScript, parseSvelte } from '../core/tooling/parsers.ts';
 import process from 'node:process';
+import { svelte } from '../core/index.ts';
 
 export function addEslintConfigPrettier(content: string): string {
 	const { ast, generateCode } = parseScript(content);
@@ -92,7 +93,8 @@ export function addToDemoPage(
 		}
 	}
 
-	imports.addNamed(script, { imports: ['resolve'], from: '$app/paths' });
+	svelte.ensureScript(ast, { language });
+	imports.addNamed(ast.instance.content, { imports: ['resolve'], from: '$app/paths' });
 
 	ast.fragment.nodes.unshift(...toFragment(`<a href={resolve('/demo/${path}')}>${path}</a>`));
 	ast.fragment.nodes.unshift();